+++ src/test/kotlin/org/kalasim/examples/GasStation.kt
@@ -3,7 +3,6 @@ package org.kalasim.examples
 import org.apache.commons.math3.distribution.UniformRealDistribution
 import org.kalasim.*
 import org.kalasim.misc.printThis
-import org.koin.core.component.get
 import org.koin.core.component.inject
 import org.koin.core.qualifier.named
 
@@ -100,8 +99,8 @@ object GasStation {
             }
 
             get<GasStation>().requesters.apply {
-                queueLengthMonitor.printThis()
-                lengthOfStayMonitor.printThis()
+                queueLengthTimeline.printThis()
+                lengthOfStayTimeline.printThis()
             }
         }
     }