+++ src/test/kotlin/org/kalasim/examples/bank/oneclerk/Bank1clerk.kt
@@ -65,7 +65,7 @@ fun main() {
     waitingLine.stats.print()
 
     if(canDisplay()) {
-        waitingLine.queueLengthMonitor.display()
-        waitingLine.lengthOfStayMonitor.display()
+        waitingLine.queueLengthTimeline.display()
+        waitingLine.lengthOfStayTimeline.display()
     }
 }