+++ src/test/kotlin/org/kalasim/examples/elevator/Elevator.kt
@@ -248,9 +248,9 @@ fun main() {
                 println(
                     "%5d%5d%15.3f%15.3f".format(
                         it.level,
-                        it.queue.lengthOfStayMonitor.statistics().n,
-                        it.queue.queueLengthMonitor.statistics().mean,
-                        it.queue.lengthOfStayMonitor.statistics().mean
+                        it.queue.lengthOfStayTimeline.statistics().n,
+                        it.queue.queueLengthTimeline.statistics().mean,
+                        it.queue.lengthOfStayTimeline.statistics().mean
                     )
                 )
             }