+++ src/test/kotlin/org/kalasim/test/DemoTests.kt
@@ -22,7 +22,7 @@ fun main() {
     val mm1Queue = MM1Queue().apply {
         run(100)
         server.claimedTimeline.display().show()
-        server.requesters.lengthOfStayMonitor.display().show()
+        server.requesters.lengthOfStayTimeline.display().show()
     }
 
     mm1Queue.traces.filterIsInstance<InteractionEvent>().asDataFrame().print()