+++ docs/userguide/docs/changes.md
@@ -11,6 +11,7 @@ Not yet released
 
 Minor enhancements
 
+* Added `capacity` to [component collections](collections.md)
 * Reworked [distribution support API](basics.md#randomness--distributions) for better API experience to enable controlled randomization in process models
 * [#38](https://github.com/holgerbrandl/kalasim/issues/38) Better API support for [depletable resources](resource.md#depletable-resources)